Bug fix and feature change
Added anti-aliasing to MapView

Added feature such that when user adds an obstacle to the map, the "add
obstacle" button is unchecked.

Increased size of obstacle 2x.

Added code to prevent adding of obstacle on a "right click".

Fixed bug in path generation
Problem: Path leaves artifacts, and is not generated completely
Root cause: PathObject is not centered correctly. Therefore, lines
painted are not within the bounding rect
Solution: Fixed computation of PathObject's center to be the center of
the line connecting (left, bottom) and (top, right), which are the x's
and y's representing the extremes of the set of points in the path. They
are the values used to compute the bounding rect. This solution makes it
so that the path points are within the bounding rect.